+++ Source/WebKit/Platform/IPC/MessageSender.cpp
@@ -30,15 +30,18 @@ namespace IPC {
 
 MessageSender::~MessageSender() = default;
 
-bool MessageSender::sendMessage(UniqueRef<Encoder>&& encoder, OptionSet<SendOption> sendOptions, std::optional<std::pair<CompletionHandler<void(IPC::Decoder*)>, uint64_t>>&& asyncReplyInfo)
+bool MessageSender::sendMessage(UniqueRef<Encoder>&& encoder, OptionSet<SendOption> sendOptions)
 {
     auto* connection = messageSenderConnection();
     ASSERT(connection);
-
-    if (asyncReplyInfo)
-        IPC::addAsyncReplyHandler(*connection, asyncReplyInfo->second, WTFMove(asyncReplyInfo->first));
-
     return connection->sendMessage(WTFMove(encoder), sendOptions);
+}
+
+bool MessageSender::sendMessageWithAsyncReply(UniqueRef<Encoder>&& encoder, AsyncReplyHandler replyHandler, OptionSet<SendOption> sendOptions)
+{
+    auto* connection = messageSenderConnection();
+    ASSERT(connection);
+    return connection->sendMessageWithAsyncReply(WTFMove(encoder), WTFMove(replyHandler), sendOptions);
 }
 
 } // namespace IPC
